+++ idea/testData/codeInsight/outOfBlock/InSuperTypeCallInLambaInBody.kt
@@ -1,0 +1,5 @@
+// FALSE
+
+open class A(a: () -> Unit)
+
+class B: A({ "1"<caret> })